lexical: condense files
